+++ apps/pie-storybook/CHANGELOG.md
@@ -1,5 +1,85 @@
 # Changelog
 
+## 1.40.1
+
+### Patch Changes
+
+- Updated dependencies [[`e118da0`](https://github.com/justeattakeaway/pie/commit/e118da0ff66420635f3b5e3cad1cff5884f47632)]:
+  - @justeattakeaway/pie-icons-webc@1.13.0
+  - @justeattakeaway/pie-assistive-text@0.10.8
+  - @justeattakeaway/pie-breadcrumb@0.6.3
+  - @justeattakeaway/pie-chip@0.12.8
+  - @justeattakeaway/pie-icon-button@2.0.3
+  - @justeattakeaway/pie-link@1.2.7
+  - @justeattakeaway/pie-modal@1.7.10
+  - @justeattakeaway/pie-notification@0.14.8
+  - @justeattakeaway/pie-switch@2.0.4
+  - @justeattakeaway/pie-text-input@0.28.8
+  - @justeattakeaway/pie-toast@0.10.1
+  - @justeattakeaway/pie-checkbox@0.16.8
+  - @justeattakeaway/pie-checkbox-group@0.9.8
+  - @justeattakeaway/pie-radio-group@0.9.10
+  - @justeattakeaway/pie-select@0.6.8
+  - @justeattakeaway/pie-textarea@0.16.8
+  - @justeattakeaway/pie-cookie-banner@1.6.11
+  - @justeattakeaway/pie-toast-provider@0.6.1
+
+## 1.40.0
+
+### Minor Changes
+
+- [Added] - support custom rendering positions for the toast provider ([#2435](https://github.com/justeattakeaway/pie/pull/2435)) by [@raoufswe](https://github.com/raoufswe)
+
+- creating the pie-avatar web component ([#2449](https://github.com/justeattakeaway/pie/pull/2449)) by [@KatarinaNeskovic](https://github.com/KatarinaNeskovic)
+
+### Patch Changes
+
+- [Added] - Percy setup guide ([#2450](https://github.com/justeattakeaway/pie/pull/2450)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
+- [Added] - add missing props documentation ([#2446](https://github.com/justeattakeaway/pie/pull/2446)) by [@raoufswe](https://github.com/raoufswe)
+
+- Updated dependencies [[`a790653`](https://github.com/justeattakeaway/pie/commit/a790653fabfa9301ea4d0cb89899ee50ae10ffab), [`a790653`](https://github.com/justeattakeaway/pie/commit/a790653fabfa9301ea4d0cb89899ee50ae10ffab), [`5fd79d4`](https://github.com/justeattakeaway/pie/commit/5fd79d4396b3170d99b20710c97580cee8744cbe), [`024074c`](https://github.com/justeattakeaway/pie/commit/024074cbb428968ffee71c39f7f94cb8828f5156), [`2327fff`](https://github.com/justeattakeaway/pie/commit/2327fffa4e335e7807308a0ca5e11e76f3cf0394), [`a790653`](https://github.com/justeattakeaway/pie/commit/a790653fabfa9301ea4d0cb89899ee50ae10ffab)]:
+  - @justeattakeaway/pie-toast@0.10.0
+  - @justeattakeaway/pie-toast-provider@0.6.0
+  - @justeattakeaway/pie-button@1.6.6
+  - @justeattakeaway/pie-avatar@0.1.0
+  - @justeattakeaway/pie-css@0.17.0
+  - @justeattakeaway/pie-radio-group@0.9.9
+  - @justeattakeaway/pie-cookie-banner@1.6.10
+  - @justeattakeaway/pie-modal@1.7.9
+
+## 1.39.3
+
+### Patch Changes
+
+- Updated dependencies [[`d63ca7c`](https://github.com/justeattakeaway/pie/commit/d63ca7ccb1a40619f0118bc436550c6c02f70309)]:
+  - @justeattakeaway/pie-assistive-text@0.10.7
+  - @justeattakeaway/pie-checkbox-group@0.9.7
+  - @justeattakeaway/pie-toast-provider@0.5.7
+  - @justeattakeaway/pie-cookie-banner@1.6.9
+  - @justeattakeaway/pie-lottie-player@0.2.4
+  - @justeattakeaway/pie-notification@0.14.7
+  - @justeattakeaway/pie-icon-button@2.0.2
+  - @justeattakeaway/pie-radio-group@0.9.8
+  - @justeattakeaway/pie-breadcrumb@0.6.2
+  - @justeattakeaway/pie-form-label@0.16.4
+  - @justeattakeaway/pie-text-input@0.28.7
+  - @justeattakeaway/pie-thumbnail@0.8.4
+  - @justeattakeaway/pie-checkbox@0.16.7
+  - @justeattakeaway/pie-textarea@0.16.7
+  - @justeattakeaway/pie-divider@1.4.5
+  - @justeattakeaway/pie-spinner@1.2.4
+  - @justeattakeaway/pie-button@1.6.5
+  - @justeattakeaway/pie-select@0.6.7
+  - @justeattakeaway/pie-switch@2.0.3
+  - @justeattakeaway/pie-modal@1.7.8
+  - @justeattakeaway/pie-radio@0.11.4
+  - @justeattakeaway/pie-toast@0.9.7
+  - @justeattakeaway/pie-card@0.25.5
+  - @justeattakeaway/pie-chip@0.12.7
+  - @justeattakeaway/pie-link@1.2.6
+  - @justeattakeaway/pie-tag@0.17.4
+
 ## 1.39.2
 
 ### Patch Changes