+++ packages/components/pie-assistive-text/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @justeattakeaway/pie-assistive-text
 
+## 0.10.8
+
+### Patch Changes
+
+- Updated dependencies [[`e118da0`](https://github.com/justeattakeaway/pie/commit/e118da0ff66420635f3b5e3cad1cff5884f47632)]:
+  - @justeattakeaway/pie-icons-webc@1.13.0
+
+## 0.10.7
+
+### Patch Changes
+
+- [Changed] - Add consistent git info and storybook url to package.json file ([#2443](https://github.com/justeattakeaway/pie/pull/2443)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
 ## 0.10.6
 
 ### Patch Changes