+++ packages/components/pie-button/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.6.6
+
+### Patch Changes
+
+- [Added] - add missing props documentation ([#2446](https://github.com/justeattakeaway/pie/pull/2446)) by [@raoufswe](https://github.com/raoufswe)
+
+## 1.6.5
+
+### Patch Changes
+
+- [Changed] - Add consistent git info and storybook url to package.json file ([#2443](https://github.com/justeattakeaway/pie/pull/2443)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
+- Updated dependencies [[`d63ca7c`](https://github.com/justeattakeaway/pie/commit/d63ca7ccb1a40619f0118bc436550c6c02f70309)]:
+  - @justeattakeaway/pie-spinner@1.2.4
+
 ## 1.6.4
 
 ### Patch Changes