+++ packages/components/pie-spinner/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @justeattakeaway/pie-spinner
 
+## 1.2.4
+
+### Patch Changes
+
+- [Changed] - Add consistent git info and storybook url to package.json file ([#2443](https://github.com/justeattakeaway/pie/pull/2443)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
 ## 1.2.3
 
 ### Patch Changes